@@ -19,11 +19,10 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "symfony/validator": "^5.0"
+        "php": ">=8.0",
+        "symfony/validator": "^6.0"
     },
     "require-dev": {
-        "doctrine/annotations": "^1.6",
         "bamarni/composer-bin-plugin": "^1.3",
         "captainhook/plugin-composer": "^5.0",
         "phpunit/phpunit": "^9.0",
